ELMIS-911: select the product category from the program product table where the setting is applied. Change of implementation to handle the work in Java instead of SQL.

+++ modules/core/src/main/java/org/openlmis/core/domain/ProgramProduct.java
@@ -37,7 +37,7 @@ public class ProgramProduct extends BaseModel implements Importable {
   @ImportField(name = "Is Active", type = "boolean", mandatory = true)
   private boolean active;
 
-  private Integer productCategoryId;
+  private Long productCategoryId;
 
   private Integer displayOrder;
